+++ src/test/java/ru/yandex/practicum/sleeptracker/BadSessionsCountFunctionTest.java
@@ -10,8 +10,8 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class BadSessionsCountFunctionTest {
-    private static final int ZERO = 0;
-    private static final int ONE = 1;
+    private static final long ZERO = 0L;
+    private static final long ONE = 1L;
 
     @Test
     void should_ReturnZeroForEmptyList() {
