+++ src/test/java/ru/yandex/practicum/sleeptracker/SleeplessNightsFunctionTest.java
@@ -11,8 +11,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class SleeplessNightsFunctionTest {
 
-    private static final int ZERO = 0;
-    private static final int ONE_DAY = 1;
+    private static final long ZERO = 0L;
+    private static final long ONE_DAY = 1L;
 
     @Test
     void should_ReturnZeroForEmptyList() {
